staging: android: lmk: check free memory when tasks switch to background
Add more free-memory check point.
This patch trigger low memory killer to scan free memory when a process is moved
to background task groups.

Reviewed-on: http://gerrit.pcs.mot.com/557973
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Yi-Wei Zhao <gbjc64@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

Conflicts:
	include/linux/sched.h

Conflicts:
	drivers/staging/android/lowmemorykiller.c

+++ fs/proc/base.c
@@ -1082,7 +1082,7 @@ static ssize_t oom_score_adj_write(struct file *file, const char __user *buf,
 		goto err_sighand;
 	}
 
-	task->signal->oom_score_adj = (short)oom_score_adj;
+	task->signal->oom_score_adj = oom_score_adj;
 	delete_from_adj_tree(task);
 	add_2_adj_tree(task);
 	if (has_capability_noaudit(current, CAP_SYS_RESOURCE))
